+++ src/main/java/de/alpharogroup/regex/RegExExtensions.java
@@ -3,24 +3,20 @@
  *
  * Copyright (C) 2015 Asterios Raptis
  *
- * Permission is hereby granted, free of charge, to any person obtaining
- * a copy of this software and associated documentation files (the
- * "Software"), to deal in the Software without restriction, including
- * without limitation the rights to use, copy, modify, merge, publish,
- * distribute, sublicense, and/or sell copies of the Software, and to
- * permit persons to whom the Software is furnished to do so, subject to
- * the following conditions:
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
+ * associated documentation files (the "Software"), to deal in the Software without restriction,
+ * including without limitation the rights to use, copy, modify, merge, publish, distribute,
+ * sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
  *
- * The above copyright notice and this permission notice shall be
- * included in all copies or substantial portions of the Software.
+ * The above copyright notice and this permission notice shall be included in all copies or
+ * substantial portions of the Software.
  *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
- * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
- * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
- * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
- * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
- * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT
+ * NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+ * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 package de.alpharogroup.regex;
 
@@ -105,31 +101,38 @@ public final class RegExExtensions
 		return newQuery;
 	}
 
-    /**
-     * Checks if the given regular expression pattern is matching with the given text.
-     *
-     * @param regexPattern the regular expression pattern
-     * @param text         the text to check if it matches
-     * @return true if the given text is matching otherwise false
-     */
-    public static boolean isMatching(String regexPattern, String text) {
-        return 0 < countMatches(regexPattern, text);
-    }
+	/**
+	 * Checks if the given regular expression pattern is matching with the given text.
+	 *
+	 * @param regexPattern
+	 *            the regular expression pattern
+	 * @param text
+	 *            the text to check if it matches
+	 * @return true if the given text is matching otherwise false
+	 */
+	public static boolean isMatching(String regexPattern, String text)
+	{
+		return 0 < countMatches(regexPattern, text);
+	}
 
-    /**
-     * Count how many times the given text is matching and returns the result.
-     *
-     * @param regexPattern the regular expression pattern
-     * @param text         the text to check if it matches
-     * @return the count of how many times the given text is matching
-     */
-    public static int countMatches(String regexPattern, String text) {
-        Matcher matcher = Pattern.compile(regexPattern).matcher(text);
-        int matches = 0;
-        while (matcher.find()) {
-            matches++;
-        }
-        return matches;
-    }
+	/**
+	 * Count how many times the given text is matching and returns the result.
+	 *
+	 * @param regexPattern
+	 *            the regular expression pattern
+	 * @param text
+	 *            the text to check if it matches
+	 * @return the count of how many times the given text is matching
+	 */
+	public static int countMatches(String regexPattern, String text)
+	{
+		Matcher matcher = Pattern.compile(regexPattern).matcher(text);
+		int matches = 0;
+		while (matcher.find())
+		{
+			matches++;
+		}
+		return matches;
+	}
 
 }