+++ src/test/java/de/alpharogroup/string/StringExtensionsTest.java
@@ -3,24 +3,20 @@
  *
  * Copyright (C) 2015 Asterios Raptis
  *
- * Permission is hereby granted, free of charge, to any person obtaining
- * a copy of this software and associated documentation files (the
- * "Software"), to deal in the Software without restriction, including
- * without limitation the rights to use, copy, modify, merge, publish,
- * distribute, sublicense, and/or sell copies of the Software, and to
- * permit persons to whom the Software is furnished to do so, subject to
- * the following conditions:
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
+ * associated documentation files (the "Software"), to deal in the Software without restriction,
+ * including without limitation the rights to use, copy, modify, merge, publish, distribute,
+ * sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
  *
- * The above copyright notice and this permission notice shall be
- * included in all copies or substantial portions of the Software.
+ * The above copyright notice and this permission notice shall be included in all copies or
+ * substantial portions of the Software.
  *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
- * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
- * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
- * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
- * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
- * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT
+ * NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+ * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 package de.alpharogroup.string;
 
@@ -39,6 +35,7 @@ import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;
 
 import de.alpharogroup.BaseTestCase;
+import de.alpharogroup.collections.list.ListFactory;
 import de.alpharogroup.collections.map.MapFactory;
 import de.alpharogroup.test.objects.Person;
 
@@ -393,6 +390,29 @@ public class StringExtensionsTest extends BaseTestCase
 		expected = null;
 		actual = StringExtensions.lastCharacterToUpperCase(null);
 		assertEquals(actual, expected);
+	}
+
+	/**
+	 * Test method for {@link StringExtensions#filterStringsWithSeperator(List, String)}
+	 */
+	@Test
+	public void testPrepareForNextIteration()
+	{
+		List<String> actual;
+		List<String> expected;
+		List<String> ignoreFieldNames;
+
+		ignoreFieldNames = ListFactory.newArrayList("id", "serialVersionUID", "address.street",
+			"address.city", "contact.name", "contact.email");
+		actual = StringExtensions.filterStringsWithSeperator(ignoreFieldNames, ".");
+		expected = ListFactory.newArrayList("street", "city", "name", "email");
+		assertEquals(expected, actual);
+
+		ignoreFieldNames = ListFactory.newArrayList("id", "serialVersionUID", "address.street",
+			"address.city", "contact.name", "contact.email");
+		actual = StringExtensions.filterStringsWithSeperator(ignoreFieldNames, "contact.");
+		expected = ListFactory.newArrayList("name", "email");
+		assertEquals(expected, actual);
 	}
 
 	/**