+++ CHANGELOG.md
@@ -12,6 +12,7 @@ CHANGED:
 
 - moved class RegExExtensions from jcommons-lang project to this project
 - moved class StringOutputStream from jcommons-lang project to this project
+- update of commons-lang3 version to 3.9
 
 Version 5.2
 -------------